refactor(workspace): split keyboard shortcut definitions
